+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 [//]: # (https://keepachangelog.com/en/1.1.0/)
 
-## 
+##  [0.11.0] -
 
 - remove throwing errors from rules
-- add `unicorn/numeric-separators-style` to `recommended` config 
+- add `unicorn/numeric-separators-style` to `recommended` config
+- replace `import-x/order` with `perfectionist/sort-imports`: now sort side-effects as first group and styles as last group
 
 ## [0.10.0] - 2026-02-23